Move tests to correct location
